prevent memory leaks in components with dynamic observations
To do this we unsubscribe all observations before calling render function, which will establish new ones. This is quite fast, so the only problem is that during unsub if component was the only subscriber then subscription will be “GC'ed” by ampere and removed from the cache. It is especially bad for AJAX subscriptions which will restart every component render in such scenario. To prevent it we've wrapped unsub-all and render (= sub all) into the ampere's `dosync` macro and enchanced the latter one to batch not only computing dirty sinks, but GC'ing dirty sources too. Modest refactor of names in `freactive.core` also happened to make code clearer.